supplier_evaluation_tool/requirements.txt
streamlit==1.24.0
scikit-learn==1.3.0
matplotlib==3.7.2
google-generativeai==0.3.1
fpdf==1.7.2
pandas==2.0.3
numpy==1.25.2